pyproject.toml
[tool.poetry]
name = "constellations"
version = "0.1.0"
description = ""
authors = ["marie <marie@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^5.3.5"
pytest-cov = "^2.8.1"
autoflake = "^1.3.1"
isort = "^4.3.21"
black = "^19.10b0"
unify = "^0.5"
flake8 = "^3.7.9"
pylint = "^2.4.4"
mypy = "^0.761"
environs = "^7.4.0"
pyyaml = "^5.3.1"
pathlib = "^1.0.1"
pytest-asyncio = "^0.11.0"
asynctest = "^0.13.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
